✅ test     : SSE_view 테스트코드 작성완료
